# openapi.yml
openapi: 3.0.0
info:
  title: ASAP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/organizations/switch:
    post:
      summary: Switch organization
      operationId: switchOrganization
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Switch to another organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchOrganizationRequest"
      responses:
        "200":
          description: Switch successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/available:
    get:
      summary: Find all available organizations
      operationId: findAvailableOrganizations
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Find all available organizations for the current user
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Organization name filter
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Page number (starts from 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Number of items per page
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        "200":
          description: Organizations response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAvailableOrganizationResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/organizations:
    post:
      summary: Create organization
      operationId: createAdminOrganization
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "200":
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Find all organizations
      operationId: findAdminOrganizations
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Find all organizations
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Organization name
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by active status
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Page number (starts from 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Number of items per page
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        "200":
          description: Organizations response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrganizationResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/organizations/{id}:
    get:
      summary: Get organization by ID
      operationId: getAdminOrganizationById
      tags:
        - organizations
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update organization
      operationId: updateAdminOrganization
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete organization by ID
      operationId: deleteAdminOrganizationById
      tags:
        - organizations
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      responses:
        "200":
          description: Organization deleted successfully
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ========= Organization Members =========
  /api/organizations/{id}/members:
    get:
      summary: List organization members
      operationId: listOrganizationMembers
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: List all members of an organization with pagination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search by name or email
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserRole"
          description: Filter by role
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by active status
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Page number (starts from 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Number of items per page
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        "200":
          description: Organization members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrganizationMemberResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Add organization member
      operationId: addOrganizationMember
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Add a new member to an organization by email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrganizationMemberRequest"
      responses:
        "201":
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMember"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Member already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/{id}/members/{memberId}:
    patch:
      summary: Update organization member role
      operationId: updateOrganizationMember
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Update a member's role or active status within an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Member ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationMemberRequest"
      responses:
        "200":
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMember"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Remove organization member
      operationId: removeOrganizationMember
      tags:
        - organizations
      security:
        - bearerAuth: []
      description: Remove a member from an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Member ID
      responses:
        "204":
          description: Member removed successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        twoFactorAuthToken:
          type: string
          description: Temporary token for 2FA verification (if applicable)
        twoFactorAuthRequired:
          type: boolean
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        path:
          type: string
          description: Request path that caused the error
    Timestamp:
      type: string
      format: date-time
      description: Timestamp in ISO 8601 format
    PaginatedResponse:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (zero-based)
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages

    # ========= User Schemas =========
...
    UserRole:
      type: string
      enum:
        - PLATFORM_ADMIN
        - OWNER
        - ADMIN
        - MEMBER
        - VIEWER
        - ACCOUNTANT
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Organization ID
        name:
          type: string
          description: Organization name
        slug:
          type: string
          description: URL-friendly identifier for the organization
        isActive:
          type: boolean
          description: Whether the organization is active
        settings:
          type: object
          additionalProperties: true
          description: Organization-level settings
        billingInfo:
          type: object
          additionalProperties: true
          description: Organization billing information
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    AvailableOrganizationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Organization ID
        name:
          type: string
          description: Organization name
...
    SwitchOrganizationRequest:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
          description: Organization ID to switch to
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Organization name
        slug:
          type: string
          description: URL-friendly identifier for the organization
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          nullable: true
          type: string
          description: Organization name
        slug:
          nullable: true
          type: string
          description: URL-friendly identifier for the organization
        isActive:
          nullable: true
          type: boolean
          description: Whether the organization is active
        settings:
          nullable: true
          type: object
          additionalProperties: true
          description: Organization-level settings
        billingInfo:
          nullable: true
          type: object
          additionalProperties: true
          description: Organization billing information
    PaginatedOrganizationResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Organization"
              description: List of organizations
    PaginatedAvailableOrganizationResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AvailableOrganizationResponse"
              description: List of available organizations for the user

    # ========= Organization Member Schemas =========
    OrganizationMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Membership ID
        organizationId:
          type: string
          format: uuid
          description: Organization ID
        userId:
          type: string
          format: uuid
          description: User ID
        userEmail:
          type: string
          description: User's email address
        userFirstName:
          type: string
          description: User's first name
        userLastName:
          type: string
          description: User's last name
        role:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean
          description: Whether the membership is active
        joinedAt:
          type: string
          format: date-time
          description: Timestamp when the member joined
        invitedAt:
          type: string
          format: date-time
          description: Timestamp when the member was invited
    AddOrganizationMemberRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
          description: Email address of the user to add
        role:
          $ref: "#/components/schemas/UserRole"
    UpdateOrganizationMemberRequest:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean
          description: Whether the membership is active
    PaginatedOrganizationMemberResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrganizationMember"
              description: List of organization members